Separate lib from main
